Add initial logic to schedule component
In order to make the arrows (to move around weeks) work, it was needed to be implemented the initial logic to make this work. Now there is a variable called "currentDate" in charge of keeping track of the changes in weeks. The back button is also disabled when trying to go before the current date.