add missing properties in EmptySessionModal
Empty session modal, given the new designs, is supossed to have a "link" in its data, as well as clients for the edit version. Both things were added. The link was added to the create empty session, and the clients was added to the edit empty session.
This also removes many testing console logs that were everywhere