Complete refactor of plan-related pages
This greatly improves all plan-related pages, making them much easier to work with, following the same design patterns and making them easier for the users to understand.
This greatly reduces the amount of lines of code needed for these pages and also helps separate concerns by moving types to the types directory.
It also makes use of typescript everywhere making it much easier to see errors and make refactors.